Entity ids are now int instead of long

+++ src/main/java/org/owasp/securityshepherd/model/Module.java
@@ -16,7 +16,7 @@ public final class Module {
 
 	@EqualsAndHashCode.Include
 	@Id
-	private final long id;
+	private final int id;
 
 	@NonNull
 	private final String name;
